Объекты Departament и Employee выделены в отдельные файлы
