+++ WpfApp/Employee.cs
@@ -1,0 +1,13 @@
+﻿namespace WpfApp
+{
+    public class Employee
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public int Age { get; set; }
+        public override string ToString()
+        {
+            return $"Id:{Id}, Name:{Name}, Age:{Age}";
+        }
+    }
+}
